+++ CHANGELOG.md
@@ -17,7 +17,8 @@
 ### Bugfixes
 - Resolve console.warning() is not a function (when drawing SIDs), by removing the -ing (https://github.com/openscope/openscope/issues/864)
 - Restore functionality of non-procedural descents to airspace ceiling [#879](https://github.com/openscope/openscope/issues/879)
-- Runways where aircraft have taxied to and gone to another runway do not display subsequent aircraft [#870](https://github.com/openscope/openscope/issues/870)
+- Ensure runways show rest of departure line after first a/c is changed to a different runway [#870](https://github.com/openscope/openscope/issues/870)
+